org: rename package from `torch-angular-sampling` to `torch-so3`
