+++ README.md
@@ -1,10 +1,10 @@
-# torch-angular-search
+# torch-so3
 
-[![License](https://img.shields.io/pypi/l/torch-angular-search.svg?color=green)](https://github.com/jdickerson95/torch-angular-search/raw/main/LICENSE)
-[![PyPI](https://img.shields.io/pypi/v/torch-angular-search.svg?color=green)](https://pypi.org/project/torch-angular-search)
-[![Python Version](https://img.shields.io/pypi/pyversions/torch-angular-search.svg?color=green)](https://python.org)
-[![CI](https://github.com/jdickerson95/torch-angular-search/actions/workflows/ci.yml/badge.svg)](https://github.com/jdickerson95/torch-angular-search/actions/workflows/ci.yml)
-[![codecov](https://codecov.io/gh/jdickerson95/torch-angular-search/branch/main/graph/badge.svg)](https://codecov.io/gh/jdickerson95/torch-angular-search)
+[![License](https://img.shields.io/pypi/l/torch-so3.svg?color=green)](https://github.com/teamtomo/torch-so3/raw/main/LICENSE)
+[![PyPI](https://img.shields.io/pypi/v/torch-so3.svg?color=green)](https://pypi.org/project/torch-so3)
+[![Python Version](https://img.shields.io/pypi/pyversions/torch-so3.svg?color=green)](https://python.org)
+[![CI](https://github.com/teamtomo/torch-so3/actions/workflows/ci.yml/badge.svg)](https://github.com/teamtomo/torch-so3/actions/workflows/ci.yml)
+[![codecov](https://codecov.io/gh/teamtomo/torch-so3/branch/main/graph/badge.svg)](https://codecov.io/gh/teamtomo/torch-so3)
 
 Generate uniform 3D euler angles (ZYZ)
 
@@ -12,13 +12,13 @@ Generate uniform 3D euler angles (ZYZ)
 
 Install via pip
 ```zsh
-pip install torch-angular-search
+pip install torch-so3
 ```
 
 Install via source by first cloning the repository then running.
 ```zsh
-git clone https://github.com/jdickerson95/torch-angular-search.git
-cd torch-angular-search
+git clone https://github.com/teamtomo/torch-so3.git
+cd torch-so3
 pip install -e .
 ```
 And for development and testing use
@@ -37,7 +37,7 @@ pre-commit run
 A basic example of generating uniform Euler angles in 4.0 and 6.0 degree increments across the entire SO(3) group is shown below.
 
 ```python
-from torch_angular_search.hopf_angles import get_uniform_euler_angles
+from torch_so3.hopf_angles import get_uniform_euler_angles
 
 angles = get_uniform_euler_angles(
     in_plane_step=4.0,  # units of degrees
