+++ src/torch_so3/__init__.py
@@ -3,7 +3,7 @@
 from importlib.metadata import PackageNotFoundError, version
 
 try:
-    __version__ = version("torch-angular-search")
+    __version__ = version("torch-so3")
 except PackageNotFoundError:
     __version__ = "uninstalled"
 __author__ = "Josh Dickerson"
